🚨 Extract `remote` from `models_config`
TG-2031
