+++ src/youwol/app/environment/models/__init__.py
@@ -5,6 +5,7 @@ The models used in YouWolEnvironment and its Configuration.
 # relative
 from .defaults import *
 from .flow_switches import *
+from .model_remote import *
 from .models import *
 from .models_config import *
 from .models_project import *
